+++ front/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - 2023-11-22
+
+### Added
+
+- New resource search bar UI implementation
+
 ## [0.4.1] - 2023-11-21
 
 ### Fixed
@@ -34,8 +40,19 @@ All notable changes to this project will be documented in this file.
 
 - My Resources widget in Category page
 
+## [0.3.0] - 2023-11-16
+
+## Added
+
+- New resource search bar UI implementation
+
 ## [0.2.1] - 2023-11-16
 
+
+### Added
+
+- Implement search bar resources UI.
+  
 ### Changed
 
 - Refactored votes/dates logic into a single component (VotesDateController).
@@ -44,6 +61,7 @@ All notable changes to this project will be documented in this file.
 
 - Fixed 'Votos' & 'Fecha' visual inconsistencies (arrow icon asc/desc, initial state).
 - Incorporated missing translations.
+
 
 ## [0.1.1] - 2023-11-15